Add DensityStructure to __init__; rename grid.py -> density_grid.py
