+++ src/lintsampler/__init__.py
@@ -7,9 +7,10 @@ See README file and online documentation (https://lintsampler.readthedocs.io)
 for further details and usage instructions. 
 """
 from .lintsampler import LintSampler
-from .grid import DensityGrid
-from importlib.metadata import version, PackageNotFoundError
+from .density_structure import DensityStructure
+from .density_grid import DensityGrid
+from importlib.metadata import version
 
 __version__ = version("lintsampler")
 
-__all__ = ["LintSampler", "DensityGrid"]
+__all__ = ["LintSampler", "DensityStructure", "DensityGrid"]
