+++ src/lintsampler/density_grid.py
